+++ src/c/augnet/http.h
@@ -21,14 +21,14 @@ struct aug_httphandlers {
 typedef struct aug_httpparser_* aug_httpparser_t;
 
 AUGNET_API aug_httpparser_t
-aug_createhttpparser(size_t size, const struct aug_httphandlers* handlers,
+aug_createhttpparser(unsigned size, const struct aug_httphandlers* handlers,
                      const struct aug_var* arg);
 
 AUGNET_API int
 aug_freehttpparser(aug_httpparser_t parser);
 
 AUGNET_API int
-aug_parsehttp(aug_httpparser_t parser, const char*, size_t size);
+aug_parsehttp(aug_httpparser_t parser, const char*, unsigned size);
 
 AUGNET_API int
 aug_endhttp(aug_httpparser_t parser);
